Migrate dragonfly assets to v3.x cleanly
When v2.x clients use the v3 migration script, none of their content
assets can be found because the job ID has changed as we upgraded the
Dragonfly library. When upgrading Dragonfly assets, try to find an asset
match by the filename of the image rather than the entire asset URL.
This isn't 100% perfect because we don't validate uniqueness on content
asset filenames, but it should get us most of the way there.

UPGRADE-30