Avoid Method Errors When Upgrading New Gems
When a new gem has been included in an upgrading gem's dependencies, an
error can be thrown when viewing the diff or report if the gem is not
already included in the app's Gemfile prior to running the upgrade
tools. Ensure that `Workarea::Upgrade::Diff#find_from_path!` returns
`nil` when this occurs, and the diff generator skips over gems that it
cannot produce a `Workarea::Upgrade::GemDiff` for.

UPGRADE-7