Fix incomplete Rails environment (#7)
Ran in to an issue where running this would throw an error about a missing ActionView constant.
Found reports of the same error in other rails-adjacent projects where part of the base rails
libraries weren't getting loaded or were loading in the wrong order. Explicitly pulling in
the ActionView railtie here seems to have fixed the problem.